+++ include/confetti_util.h
@@ -4,10 +4,10 @@
 struct ConfettiUtil
 {
     struct OamData oam;
-    s16 x;
-    s16 y;
-    s16 xDelta;
-    s16 yDelta;
+    s32 x;
+    s32 y;
+    s32 xDelta;
+    s32 yDelta;
     u32 tileTag;
     u32 palTag;
     u32 tileNum;
@@ -18,7 +18,7 @@ struct ConfettiUtil
     u8 allowUpdates:1;
     u8 dummied:1;
     u8 priority:2;
-    s16 data[8];
+    s32 data[8];
     void (*callback)(struct ConfettiUtil *);
 };
 
@@ -26,8 +26,8 @@ bool32 ConfettiUtil_Init(u8 count);
 bool32 ConfettiUtil_Free(void);
 bool32 ConfettiUtil_Update(void);
 u8 ConfettiUtil_SetCallback(u8 id, void (*func)(struct ConfettiUtil *));
-u8 ConfettiUtil_SetData(u8 id, u8 dataArrayId, s16 dataValue);
-u8 ConfettiUtil_AddNew(const struct OamData *oam, u32 tileTag, u32 palTag, s16 x, s16 y, u8 animNum, u8 priority);
+u8 ConfettiUtil_SetData(u8 id, u8 dataArrayId, s32 dataValue);
+u8 ConfettiUtil_AddNew(const struct OamData *oam, u32 tileTag, u32 palTag, s32 x, s32 y, u8 animNum, u8 priority);
 u8 ConfettiUtil_Remove(u8 id);
 
 #endif // GUARD_CONFETTI_UTIL_H
