+++ include/contest_link.h
@@ -14,7 +14,7 @@ void Task_LinkContest_CommunicateRound1Points(u8);
 void Task_LinkContest_CommunicateTurnOrder(u8);
 void Task_LinkContest_FinalizeConnection(u8);
 void Task_LinkContest_CommunicateMoveSelections(u8);
-bool32 LinkContest_TryLinkStandby(s16 *);
+bool32 LinkContest_TryLinkStandby(s32 *);
 bool32 LinkContest_SendBlock(void *, u32);
 bool32 LinkContest_GetBlockReceivedFromAllPlayers(void);
 bool32 LinkContest_GetBlockReceived(u8);
