+++ include/dma3.h
@@ -48,8 +48,8 @@
 
 void ClearDma3Requests(void);
 void ProcessDma3Requests(void);
-s16 RequestDma3Copy(const void *src, void *dest, u32 size, u32 mode);
-s16 RequestDma3Fill(s32 value, void *dest, u32 size, u32 mode);
-s16 CheckForSpaceForDma3Request(s16 index);
+s32 RequestDma3Copy(const void *src, void *dest, u32 size, u32 mode);
+s32 RequestDma3Fill(s32 value, void *dest, u32 size, u32 mode);
+s32 CheckForSpaceForDma3Request(s32 index);
 
 #endif // GUARD_DMA3_H
