+++ include/intro_credits_graphics.h
@@ -18,8 +18,8 @@ enum {
 };
 
 extern u32 gIntroCredits_MovingSceneryVBase;
-extern s16 gIntroCredits_MovingSceneryVOffset;
-extern s16 gIntroCredits_MovingSceneryState;
+extern s32 gIntroCredits_MovingSceneryVOffset;
+extern s32 gIntroCredits_MovingSceneryState;
 
 extern const struct CompressedSpriteSheet gSpriteSheet_IntroBrendan[];
 extern const struct CompressedSpriteSheet gSpriteSheet_IntroMay[];
@@ -39,8 +39,8 @@ void LoadCreditsSceneGraphics(u8);
 void SetCreditsSceneBgCnt(u8);
 u8 CreateBicycleBgAnimationTask(u8 mode, u32 bg1Speed, u32 bg2Speed, u32 bg3Speed);
 void CycleSceneryPalette(u8);
-u8 CreateIntroBrendanSprite(s16 x, s16 y);
-u8 CreateIntroMaySprite(s16 x, s16 y);
-u8 CreateIntroFlygonSprite(s16 x, s16 y);
+u8 CreateIntroBrendanSprite(s32 x, s32 y);
+u8 CreateIntroMaySprite(s32 x, s32 y);
+u8 CreateIntroFlygonSprite(s32 x, s32 y);
 
 #endif // GUARD_INTRO_CREDITS_GRAPHICS_H
