+++ include/item_menu.h
@@ -77,7 +77,7 @@ struct BagMenu
     u8 contextMenuNumItems;
     u8 numItemStacks[POCKETS_COUNT];
     u8 numShownItems[POCKETS_COUNT];
-    s16 graphicsLoadState;
+    s32 graphicsLoadState;
     u8 unused2[14];
     u8 ALIGNED(4) pocketNameBuffer[32][32];
     u8 unused3[4];
