+++ include/minigame_countdown.h
@@ -1,7 +1,7 @@
 #ifndef GUARD_MINIGAME_COUNTDOWN_H
 #define GUARD_MINIGAME_COUNTDOWN_H
 
-void StartMinigameCountdown(u32 tilesTag, u32 palTag, s16 x, s16 y, u8 subpriority);
+void StartMinigameCountdown(u32 tilesTag, u32 palTag, s32 x, s32 y, u8 subpriority);
 bool32 IsMinigameCountdownRunning(void);
 
 #endif //GUARD_MINIGAME_COUNTDOWN_H
